new code
i seperated the void methods from the comp so that adding new stuff to
the comp is now just a few lines of code calling some methods. This
means that the comp file is much less complex and easy to change/debug.
There is a new bug however. The led indicaters get stuck on green if the
user changes to arcade the robot switches back in 200milisecs. please
take a close look at the methods file and comp file before making
changes.